update go test
