+++ client/README
@@ -1,1 +1,1 @@
-go test -conf=./test-examples.conf
+go test -tt=./test-examples.conf
